backend/requirements.txt
# ── Core ───────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── AI ─────────────────────────────────────────────────────────────────────
openai>=1.30.0
google-generativeai>=0.5.0

# ── Config / env ────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
